+++ transmitter.py
@@ -1,38 +1,8 @@
 import  numpy as np
 import wave
 from scipy.signal import butter, lfilter
+from shared_utils import *
 import pyaudio
-
-
-# noinspection PyTupleAssignmentBalance
-def butter_bandpass(lowcut, highcut, fs, order=5):
-    """
-    this function designs a Butterworth filter
-    :param lowcut: lower cutoff frequency
-    :param highcut: upper cutoff frequency
-    :param fs: sampling frequency
-    :param order: order of filter
-    :return:  coefficients of the designed filter
-    """
-    nyquist = fs / 2            # Nyquist Theorem
-    low = lowcut / nyquist      # Normalization
-    high = highcut / nyquist    # Normalization
-    b, a = butter(order, [low, high], btype='band')
-    return b, a         # b - numerator a - denominator
-
-def apply_bandpass_filter(data, lowcut, highcut, fs, order=5):
-    """
-    this function applies Butterworth filter on a given data
-    :param data: given binary data
-    :param lowcut: lower cutoff frequency
-    :param highcut: upper cutoff frequency
-    :param fs: sampling frequency
-    :param order: order of filter
-    :return: filtered data
-    """
-    b, a = butter_bandpass(lowcut, highcut, fs, order=order)
-    return lfilter(b, a, data)
-
 
 def hamming_encode(data):
     """
@@ -57,7 +27,7 @@ def hamming_encode(data):
 
 
 # noinspection PyUnresolvedReferences
-def modulate_message(text, output_file, lowcut=400, highcut=1200, duration=0.1, sample_rate=44100):
+def modulate_message(text, output_file, lowcut=480, highcut=1020, duration=0.1, sample_rate=44100):
     """
     modulates a text message into a sound using FSK modulation filtered by BPF and encoded with Hamming code
     :param text: text to be modulated
@@ -67,19 +37,21 @@ def modulate_message(text, output_file, lowcut=400, highcut=1200, duration=0.1, 
     :param highcut: upper cutoff frequency
     :param duration: bit duration
     :param sample_rate: sampling rate
-    :return:
+    :return: the normalized modulated signal
     """
     binary_data = ''.join(format(ord(char),'08b') for char in text)
     encoded_binary_data = hamming_encode(binary_data)
+    encoded_binary_with_barkers = START_BARKER + encoded_binary_data + END_BARKER
     print(f'Original binary data: {binary_data}')
     print(f'Hamming encoded binary data: {encoded_binary_data}')
+    print(f'Full encoded binary data (with barkers): {encoded_binary_with_barkers}')
 
     freq_0 = 500    # frequency for '0'
     freq_1 = 1000   # frequency for '1'
-    amplitude = 0.5 # 1 VPP
+    amplitude = 0.8
 
     signal = []
-    for bit in encoded_binary_data:
+    for bit in encoded_binary_with_barkers:
         freq = freq_1 if bit == '1' else freq_0
         t = np.linspace(0, duration, int(sample_rate * duration), endpoint=False)
         tone = amplitude * np.sin(2 * np.pi * freq * t) # sound harmony
@@ -88,6 +60,9 @@ def modulate_message(text, output_file, lowcut=400, highcut=1200, duration=0.1, 
     signal = np.array(signal, dtype=np.float32)
     filtered_signal = apply_bandpass_filter(signal, lowcut, highcut, sample_rate, order=5)  # filter the data
 
+    # Normalize the signal
+    normalized_signal = filtered_signal / np.max(np.abs(filtered_signal))
+
     # Save the filtered sound into a file
     with wave.open(output_file, 'wb') as wf:
         wf.setnchannels(1)  # Mono
@@ -95,33 +70,29 @@ def modulate_message(text, output_file, lowcut=400, highcut=1200, duration=0.1, 
         wf.setframerate(sample_rate)
         wf.writeframes((filtered_signal * 32767).astype(np.int16).tobytes())
 
-    print(f"The modulated sound saved to {output_file}")
+    return  normalized_signal
 
-def play_audio(audio_file):
+def play_audio(signal, sample_rate=44100):
     """
     This function plays an audio file (saved after modulating the message)
-    :param audio_file: audio file to play
+    :param signal: the signal to be played
+    :param sample_rate: the sample rate to play the signal
     :return:
     """
-    chunk = 1024
-    wf = wave.open(audio_file, 'rb')
     p = pyaudio.PyAudio()
-    stream = p.open(format=p.get_format_from_width(wf.getsampwidth()),channels=wf.getnchannels(),rate=wf.getframerate(),
-                    output=True)
-    data = wf.readframes(chunk)
-
-    while data:
-        stream.write(data)
-        data = wf.readframes(chunk)
-
+    mono_channels = 1
+    stream = p.open(format=pyaudio.paFloat32,channels=mono_channels,rate=sample_rate,output=True)
+    print("Playing audio...")
+    stream.write(signal.astype(np.float32).tobytes())
     stream.stop_stream()
     stream.close()
     p.terminate()
+    print("Audio playback complete.")
 
 
 # Example Usage
 if __name__ == "__main__":
     message = 'My name is Barack'
     output_file = 'My_name_is_Barack.wav'
-    modulate_message(text=message,output_file=output_file)
-    play_audio(audio_file=output_file)
+    signal = modulate_message(text=message,output_file=output_file)
+    play_audio(signal=signal)
